+++ src/resources/init.rs
@@ -2272,6 +2272,7 @@ impl DeviceResources {
             target_format,
             sample_count,
             debug_vis_shaders: false,
+            mesh_pipelines_dirty: false,
             pipeline_cache,
             solid_pipeline,
             deform,